requirements.txt
panda
dash
dash-core-components==1.15.0
dash-html-components==1.1.2
dash-bootstrap-components
gunicorn==20.0.4
plotly
xlrd== 1.1.0
pathlib
pybase64
regex
emoji
matplotlib
numpy
datetime
dash-auth==1.4.1
dash-canvas==0.1.0
dash-cytoscape==0.2.0
dash-daq==0.5.0
dash-extensions==0.0.51
dash-renderer==1.9.0
dash-table==4.11.2
pytesseract==0.3.4
imutils
scikit-image
opencv-python-headless==4.2.0.32
beautifulsoup4==4.9.1
requests==2.23.0